requirements.txt
mcp
anthropic
python-dotenv
requests
bs4
flask
openai